requirements.txt

reflex==0.8.22
python-decouple
google-genai